Revert "refactor: add version for main modules (#10)"
This reverts commit e8ab5386aadf327d15b193ce5349fbb3876ad312.
