+++ .dockerignore
@@ -1,2 +1,2 @@
-build
+platforms
 third_party
