+++ cmd/gateway/v1/main.go
@@ -14,17 +14,16 @@
  * limitations under the License.
  */
 
-package registry
+package main
 
-import "github.com/convrz/convers/v1/core/cvzapp"
+import (
+	"log"
 
-func New() cvzapp.App {
-	return &App{}
-}
+	"github.com/convrz/convers/core/cvzfactory"
+	"github.com/convrz/convers/internal/apps/gateway/v1"
+)
 
-type App struct{}
-
-func (a *App) Run() error {
-	//TODO implement me
-	panic("implement me")
+func main() {
+	app := cvzfactory.Build(gateway.New)
+	log.Fatal(app.Run())
 }
