+++ core/cvzfactory/factory_helper.go
@@ -18,11 +18,10 @@ package cvzfactory
 
 import (
 	"context"
+	"github.com/convrz/convers/core/internal"
 	"log"
 
-	cvzinternal "github.com/convrz/convers/v1/core/internal"
-
-	"github.com/convrz/convers/v1/core/cvzapp"
+	"github.com/convrz/convers/core/cvzapp"
 
 	"go.uber.org/fx"
 )
