+++ internal/apps/registry/v1/server/registry_server.go
