+++ x/greeter/v1/cmd/main.go
@@ -19,7 +19,7 @@ package main
 import (
 	"log"
 
-	"github.com/convrz/convers/v1/core/cvzfactory"
+	"github.com/convrz/convers/core/cvzfactory"
 	"github.com/convrz/convers/x/greeter/v1/app"
 
 	_ "github.com/convrz/convers/x/greeter/v1/internal/init"
